Refactoring : Step 1
