+++ src/index.ts
@@ -3,10 +3,10 @@ import { isDebugEnabled } from "./loggerConfiguration";
 import * as xcMessages from "./communication/xcomponentMessages";
 
 export { xcLogLevels } from "./loggerConfiguration";
-export { Connection } from "./communication/xcConnection";
-export { Session } from "./communication/xcSession";
-export { Publisher } from "./communication/xcWebSocketPublisher";
-export { Subscriber } from "./communication/xcWebSocketSubscriber";
+export { Connection } from "./interfaces/Connection";
+export { Session } from "./interfaces/Session";
+export { Publisher } from "./interfaces/Publisher";
+export { Subscriber } from "./interfaces/Subscriber";
 export { xcMessages };
 
 export default new xcomponentapi();
