+++ src/interfaces/Publisher.ts
@@ -1,0 +1,9 @@
+import { ApiConfiguration } from "../configuration/apiConfiguration";
+
+export interface Publisher {
+    configuration: ApiConfiguration;
+    privateTopic: string;
+    sendWithStateMachineRef(stateMachineRef: any, messageType: string, jsonMessage: any, visibilityPrivate: boolean, specifiedPrivateTopic: string): void;
+    send(componentName: string, stateMachineName: string, messageType: string, jsonMessage: any, visibilityPrivate: boolean, specifiedPrivateTopic: string): void;
+    canPublish(componentName: string, stateMachineName: string, messageType: string): boolean;
+}
